refactor: Image를 도메인 패키지로 변경한다
