+++ src/main/java/daybyquest/badge/application/SaveBadgeService.java
@@ -3,9 +3,9 @@ package daybyquest.badge.application;
 import daybyquest.badge.domain.Badge;
 import daybyquest.badge.domain.Badges;
 import daybyquest.global.utils.MultipartFileUtils;
-import daybyquest.image.vo.Image;
-import daybyquest.image.vo.ImageIdentifierGenerator;
-import daybyquest.image.vo.Images;
+import daybyquest.image.domain.Image;
+import daybyquest.image.domain.ImageIdentifierGenerator;
+import daybyquest.image.domain.Images;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
