+++ src/main/java/daybyquest/badge/query/BadgeData.java
@@ -1,6 +1,6 @@
 package daybyquest.badge.query;
 
-import daybyquest.image.vo.Image;
+import daybyquest.image.domain.Image;
 import java.time.LocalDateTime;
 import lombok.Getter;
