+++ src/main/java/daybyquest/group/application/SaveGroupService.java
@@ -4,9 +4,9 @@ import daybyquest.global.utils.MultipartFileUtils;
 import daybyquest.group.domain.Group;
 import daybyquest.group.domain.Groups;
 import daybyquest.group.dto.request.SaveGroupRequest;
-import daybyquest.image.vo.Image;
-import daybyquest.image.vo.ImageIdentifierGenerator;
-import daybyquest.image.vo.Images;
+import daybyquest.image.domain.Image;
+import daybyquest.image.domain.ImageIdentifierGenerator;
+import daybyquest.image.domain.Images;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
