+++ src/main/java/daybyquest/image/domain/Image.java
@@ -1,4 +1,4 @@
-package daybyquest.image.vo;
+package daybyquest.image.domain;
 
 import static lombok.AccessLevel.PROTECTED;
