+++ src/main/java/daybyquest/interest/application/SaveInterestService.java
@@ -1,9 +1,9 @@
 package daybyquest.interest.application;
 
 import daybyquest.global.utils.MultipartFileUtils;
-import daybyquest.image.vo.Image;
-import daybyquest.image.vo.ImageIdentifierGenerator;
-import daybyquest.image.vo.Images;
+import daybyquest.image.domain.Image;
+import daybyquest.image.domain.ImageIdentifierGenerator;
+import daybyquest.image.domain.Images;
 import daybyquest.interest.domain.Interest;
 import daybyquest.interest.domain.Interests;
 import daybyquest.interest.dto.request.SaveInterestRequest;
