+++ src/main/java/daybyquest/interest/domain/Interest.java
@@ -1,7 +1,7 @@
 package daybyquest.interest.domain;
 
 import daybyquest.global.error.exception.InvalidDomainException;
-import daybyquest.image.vo.Image;
+import daybyquest.image.domain.Image;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
