+++ src/main/java/daybyquest/user/application/SaveUserService.java
@@ -1,7 +1,7 @@
 package daybyquest.user.application;
 
-import daybyquest.image.vo.BaseImageProperties;
-import daybyquest.image.vo.Image;
+import daybyquest.image.domain.BaseImageProperties;
+import daybyquest.image.domain.Image;
 import daybyquest.user.domain.User;
 import daybyquest.user.domain.Users;
 import daybyquest.user.dto.request.SaveUserRequest;
